update email login page UI
